requirements.txt
azure-functions==1.17.0
azure-storage-blob==12.19.0
azure-search-documents==11.4.0
azure-identity==1.15.0
azure-core==1.29.5
openai==1.3.5
aiohttp==3.9.1
python-multipart==0.0.6
typing-extensions==4.8.0
tenacity==8.2.3
azure-ai-formrecognizer==3.3.0
tiktoken==0.5.2
PyPDF2>=3.0.0